refactor(state): move spawn related logic out of state/mod.rs into spawn.rs
